November 5 -- Setting Pump and Button Pins, Defining Drink Functions
I re-organized the code so it is easier to look at (drinks together, buttons together, LEDs together). I also declared the pin numbers as const ints instead of using #define (if this does not work with the relay board, I will change it again later). I set pin numbers for the pumps and buttons and declared button states. I will set pin numbers for LEDs later. I also created if and else if statements for each button state, and declared seven different functions to mix the different drinks. I will add LED effects to these functions later.

The pumps I purchased are not working, so I cannot check if this code works. It compiles, so that's a good start.