November 8, 2017 -- Changed Pin Numbers, Changed If Else, Added Serial.print, Removed Vodka Tonic Option
Changed pin numbers so no relay pins are in pin 1. This means I also eliminated one of the beverage options (Vodka Tonic is no longer an option; all mentions of Vodka Tonic or VT are currently commented in case I change my mind about this later).

I added Serial.print commands to display my button states and a function to set all button states back to HIGH after the drink is made.

I changed my else if  statements to include states for all buttons in each statement, not just the button I am pushing.

The button attached to pin 13 is the only one that currently works correctly.  I am not sure why this is. I can hook up any of the wired buttons to pin 13 and it will work, but it will not work if it is attached to any of the other pin. I can also call any function when the button in pin 13 is pressed and it will work. All of my buttons and functions work, but only when they are using pin 13.