November 16 -- Fixed Buttons, Added LEDs
I set the button pins to pull down buttons without needing to use resistors in the circuit (using digitalWrite, HIGH). I also changed my if statements, so they actually work properly, by changing the syntax to include &&. Now all of my buttons work properly!

I also initialized LED pins (one LED for each pump/beverage). I set the LEDs to turn on when the cooresponding pump is on.

All of the code currently works :) All that's left to do is create the LED strobe pattern that will loop when a drink is not being made.

I have not physically attached the pumps yet to make sure they work, but the correct switches on the relay module turn on when the buttons are pressed.