Renamed Pumps and LEDs; Change Functions
Changed the names of the pumps, buttons, and LEDs to reflect the new alcohol/mixes.
Edited if statements to use the new button names and new functions.
Edited functions to only use one pump and one LED at a time, and changed the amount of time they stay on for,

Deleted commented blocks that were not being used from the original code.